+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
 
+### Added
+
+- Search integration
+
 ## [1.0.0]
 
 ### Added